pyproject.toml
[tool.poetry]
name = "ia-senac-uc14-frame-generation-project"
version = "0.1.0"
description = "Frame generation project using deep learning"
authors = ["Pablo Henrique <35077005+pablohls@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
# Deep Learning
torch = ">=2.0.0,<3.0.0"
torchvision = ">=0.15.0,<1.0.0"
# Hugging Face
transformers = ">=4.35.0"
diffusers = ">=0.24.0"
accelerate = ">=0.25.0"
# Video Processing
opencv-python = ">=4.8.0"
imageio = ">=2.31.0"
imageio-ffmpeg = ">=0.4.9"
# Image Processing
pillow = ">=10.0.0"
numpy = ">=1.24.0"
scikit-image = ">=0.21.0"
# API
fastapi = ">=0.104.0"
uvicorn = {version = ">=0.24.0", extras = ["standard"]}
python-multipart = ">=0.0.6"
aiofiles = ">=23.2.1"
# Configuration
pyyaml = ">=6.0.1"
python-dotenv = ">=1.0.0"
# Utilities
tqdm = ">=4.66.0"
loguru = ">=0.7.2"
# Optional: Interface
gradio = ">=4.0.0"
# Optional: Monitoring
tensorboard = ">=2.15.0"
wandb = ">=0.16.0"
# Jupyter
jupyter = ">=1.0.0"
ipykernel = ">=6.26.0"
matplotlib = ">=3.8.0"
lpips = "^0.1.4"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.4.0"
pytest-cov = ">=4.1.0"
black = ">=23.11.0"
flake8 = ">=6.1.0"
mypy = ">=1.7.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
